Refactor: move function
