+++ src/main/kotlin/com/github/izhangzhihao/rainbow/brackets/action/AbstractScopeHighlightingAction.kt
@@ -12,7 +12,6 @@ import com.intellij.openapi.util.Key
 import com.intellij.psi.PsiDocumentManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
-import java.awt.Color
 import java.awt.event.FocusEvent
 import java.awt.event.FocusListener
 import java.awt.event.KeyAdapter
@@ -92,16 +91,6 @@ abstract class AbstractScopeHighlightingAction : AnAction() {
             }
 
             return null
-        }
-
-        fun Color.alphaBlend(background: Color, alpha: Float): Color {
-            require(alpha in 0.0..1.0) { "alpha(0.0 <= alpha <= 1.0): $alpha" }
-
-            val r = (1 - alpha) * background.red + alpha * red
-            val g = (1 - alpha) * background.green + alpha * green
-            val b = (1 - alpha) * background.blue + alpha * blue
-
-            return Color(r.toInt(), g.toInt(), b.toInt())
         }
     }
 }
