+++ src/main/kotlin/com/github/izhangzhihao/rainbow/brackets/action/ScopeOutsideHighlightingRestrainAction.kt
@@ -2,6 +2,7 @@ package com.github.izhangzhihao.rainbow.brackets.action
 
 import com.github.izhangzhihao.rainbow.brackets.RainbowInfo
 import com.github.izhangzhihao.rainbow.brackets.settings.RainbowSettings
+import com.github.izhangzhihao.rainbow.brackets.util.alphaBlend
 import com.intellij.codeInsight.highlighting.HighlightManager
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.editor.colors.EditorColorsManager
